damage and respawn

+++ PlayerWeapon.cs
@@ -4,7 +4,7 @@
 public class PlayerWeapon
 {
     public string name = "Glock";
-    public float damage = 10.0f;
+    public int damage = 10;
     public float range = 100.0f;
 
 }
